fix functional tests
