+++ rm/rm-server/src/test/java/functionaltests/utils/RMStarterForFunctionalTest.java
@@ -25,9 +25,17 @@
  */
 package functionaltests.utils;
 
+import java.io.IOException;
+import java.security.GeneralSecurityException;
+import java.util.concurrent.ExecutionException;
+
+import org.apache.commons.configuration2.ex.ConfigurationException;
+import org.objectweb.proactive.core.config.CentralPAPropertyRepository;
 import org.ow2.proactive.resourcemanager.RMFactory;
 import org.ow2.proactive.resourcemanager.core.properties.PAResourceManagerProperties;
 import org.ow2.proactive.resourcemanager.frontend.RMConnection;
+import org.ow2.proactive.scheduler.core.properties.PASchedulerProperties;
+import org.ow2.tests.ProActiveTest;
 
 
 /**
@@ -35,6 +43,8 @@ import org.ow2.proactive.resourcemanager.frontend.RMConnection;
  *
  */
 public class RMStarterForFunctionalTest {
+
+    private static final String IAM_LOGIN_METHOD = "IAMLoginMethod";
 
     /**
      * Start a Resource Manager.
@@ -45,6 +55,8 @@ public class RMStarterForFunctionalTest {
         if (args.length < 1) {
             throw new IllegalArgumentException("RMTStarter must be started with one parameter: path to a RM Properties file");
         }
+
+        startIAMIfNeeded();
 
         String RMPropPath = args[0];
         PAResourceManagerProperties.updateProperties(RMPropPath);
@@ -57,5 +69,27 @@ public class RMStarterForFunctionalTest {
         RMConnection.waitAndJoin(null);
 
         System.out.println("Resource Manager successfully created !");
+    }
+
+    /**
+     * Start IAM microservice if it is required for authentication
+     *
+     * @throws IOException
+     * @throws InterruptedException
+     * @throws ExecutionException
+     * @throws ConfigurationException
+     */
+    public static void startIAMIfNeeded() throws IOException, InterruptedException, ExecutionException,
+            ConfigurationException, GeneralSecurityException {
+
+        //Check if PA is configured to use IAM microservice for authentication
+        if (PAResourceManagerProperties.RM_LOGIN_METHOD.getValueAsString().equals(IAM_LOGIN_METHOD)) {
+
+            String proactiveHome = CentralPAPropertyRepository.PA_HOME.getValue();
+            String bootMicroservicesPath = PASchedulerProperties.getAbsolutePath(PASchedulerProperties.SCHEDULER_BOOT_MICROSERVICES_PATH.getValueAsString());
+            String bootConfigurationPath = PASchedulerProperties.getAbsolutePath(PASchedulerProperties.SCHEDULER_BOOT_CONFIGURATION_PATH.getValueAsString());
+
+            IAMTHelper.startIAM(proactiveHome, bootMicroservicesPath, bootConfigurationPath);
+        }
     }
 }
