+++ scheduler/scheduler-server/src/test/java/functionaltests/rm/TestRMProxyRebind.java
@@ -97,8 +97,6 @@ public class TestRMProxyRebind extends MultipleRMTBase {
             TestScheduler.kill();
         }
 
-        startIAMIfNeeded();
-
         initConfigs();
     }
