+++ scheduler/scheduler-server/src/test/java/functionaltests/utils/SchedulerFunctionalTest.java
@@ -27,24 +27,18 @@ package functionaltests.utils;
 
 import static org.junit.Assert.fail;
 
-import java.io.IOException;
-import java.security.GeneralSecurityException;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 
-import org.apache.commons.configuration2.ex.ConfigurationException;
 import org.apache.log4j.Logger;
 import org.apache.log4j.PropertyConfigurator;
 import org.junit.After;
 import org.junit.Rule;
 import org.junit.rules.Timeout;
-import org.objectweb.proactive.core.config.CentralPAPropertyRepository;
 import org.ow2.proactive.resourcemanager.frontend.ResourceManager;
 import org.ow2.proactive.scheduler.common.exception.JobCreationException;
 import org.ow2.proactive.scheduler.common.job.Job;
-import org.ow2.proactive.scheduler.core.properties.PASchedulerProperties;
 import org.ow2.tests.ProActiveTest;
 
 
@@ -62,8 +56,6 @@ public class SchedulerFunctionalTest extends ProActiveTest {
     }
 
     protected static final Logger logger = Logger.getLogger("SchedulerTests");
-
-    private static final String IAM_LOGIN_METHOD = "SchedulerIAMLoginMethod";
 
     protected static SchedulerTHelper schedulerHelper;
 
@@ -120,27 +112,4 @@ public class SchedulerFunctionalTest extends ProActiveTest {
             }
         }
     }
-
-    /**
-     * Start IAM microservice if it is required for authentication
-     *
-     * @throws IOException
-     * @throws InterruptedException
-     * @throws ExecutionException
-     * @throws ConfigurationException
-     */
-    public static void startIAMIfNeeded() throws IOException, InterruptedException, ExecutionException,
-            ConfigurationException, GeneralSecurityException {
-
-        //Check if PA is configured to use IAM microservice for authentication
-        if (PASchedulerProperties.SCHEDULER_LOGIN_METHOD.getValueAsString().equals(IAM_LOGIN_METHOD)) {
-
-            String proactiveHome = CentralPAPropertyRepository.PA_HOME.getValue();
-            String bootMicroservicesPath = PASchedulerProperties.getAbsolutePath(PASchedulerProperties.SCHEDULER_BOOT_MICROSERVICES_PATH.getValueAsString());
-            String bootConfigurationPath = PASchedulerProperties.getAbsolutePath(PASchedulerProperties.SCHEDULER_BOOT_CONFIGURATION_PATH.getValueAsString());
-
-            IAMTHelper.startIAM(proactiveHome, bootMicroservicesPath, bootConfigurationPath);
-        }
-    }
-
 }
