+++ scheduler/scheduler-server/src/test/java/functionaltests/utils/SchedulerFunctionalTestNonForkModeWithRestart.java
@@ -43,8 +43,6 @@ public class SchedulerFunctionalTestNonForkModeWithRestart extends SchedulerFunc
     @BeforeClass
     public static void startSchedulerInAnyCase() throws Exception {
 
-        startIAMIfNeeded();
-
         schedulerHelper.log("Start Scheduler in non-fork mode.");
         schedulerHelper = new SchedulerTHelper(false,
                                                new File(SchedulerFunctionalTest.class.getResource("/functionaltests/config/scheduler-nonforkedscripttasks.ini")
