+++ scheduler/scheduler-server/src/test/java/functionaltests/utils/SchedulerFunctionalTestWithRestart.java
@@ -27,8 +27,6 @@ package functionaltests.utils;
 
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
-import org.objectweb.proactive.core.config.CentralPAPropertyRepository;
-import org.ow2.proactive.scheduler.core.properties.PASchedulerProperties;
 
 
 /**
@@ -42,9 +40,6 @@ public class SchedulerFunctionalTestWithRestart extends SchedulerFunctionalTest 
 
     @BeforeClass
     public static void startSchedulerInAnyCase() throws Exception {
-
-        startIAMIfNeeded();
-
         schedulerHelper.log("Starting a clean scheduler.");
         schedulerHelper = new SchedulerTHelper(true);
     }
