+++ scheduler/scheduler-server/src/test/java/functionaltests/utils/SchedulerStartForFunctionalTest.java
@@ -38,7 +38,6 @@ import org.ow2.proactive.resourcemanager.authentication.RMAuthentication;
 import org.ow2.proactive.resourcemanager.core.properties.PAResourceManagerProperties;
 import org.ow2.proactive.resourcemanager.frontend.RMConnection;
 import org.ow2.proactive.resourcemanager.frontend.ResourceManager;
-import org.ow2.proactive.resourcemanager.nodesource.NodeSource;
 import org.ow2.proactive.resourcemanager.nodesource.infrastructure.LocalInfrastructure;
 import org.ow2.proactive.resourcemanager.nodesource.policy.StaticPolicy;
 import org.ow2.proactive.scheduler.SchedulerFactory;
@@ -165,5 +164,4 @@ public class SchedulerStartForFunctionalTest implements Serializable {
 
         SchedulerConnection.waitAndJoin(schedulerUrl);
     }
-
 }
